TulipCC Tab5: RGB565 console colours, and formalise the readline UTF-8 IME patch
mphalport.c: the Tab5 console stdout draws through display_tfb_str() with
tfb_fg_pal_color / tfb_bg_pal_color, which became 16-bit (tulip_px_t) when
the Tab5 display planes went native RGB565. The extern here still declared
them uint8_t, so it read only the low byte of the transparent key ALPHA
(0x4daa -> 0xaa), which is not ALPHA, and every REPL line drew an opaque box
over the wallpaper. Widen the externs and the display_tfb_str prototype to
uint16_t under the existing TAB5 ifdef; other boards keep RGB332/uint8_t.

readline.c: commit the TULIP_READLINE_UTF8 patch
(tulip/shared/patches/micropython-readline-utf8.patch) that had been carried
only in the working tree. A multibyte character is held until complete and a
line containing any byte over 0x7f is redrawn whole, so the Tulip IME can type
Japanese at the prompt; a pure-ASCII line takes the original byte path
unchanged.

+++ ports/esp32/mphalport.c
@@ -49,9 +49,14 @@
 #include "uart.h"
 
 #ifdef TAB5
-extern uint8_t tfb_fg_pal_color;
-extern uint8_t tfb_bg_pal_color;
-void display_tfb_str(unsigned char *str, uint16_t len, uint8_t format, uint8_t fg_color, uint8_t bg_color);
+// tulip_px_t on the Tab5 -- the BG plane and the TFB colours are native RGB565,
+// so these defaults hold a 16-bit pixel (transparent is ALPHA, 0x4daa). Reading
+// them through a uint8_t extern took only the low byte (0xaa), which is not
+// ALPHA, so every console line drew an opaque box instead of letting the
+// wallpaper through. Match the real type. (Other boards keep RGB332/uint8_t.)
+extern uint16_t tfb_fg_pal_color;
+extern uint16_t tfb_bg_pal_color;
+void display_tfb_str(unsigned char *str, uint16_t len, uint8_t format, uint16_t fg_color, uint16_t bg_color);
 #endif
 
 #if MICROPY_PY_STRING_TX_GIL_THRESHOLD < 0
